Move PaymentMethod into its own folder
