--- src/Backend/Modules/CommerceMollie/Domain/Mollie/Command/UpdateMolliePaymentMethod.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Backend\Modules\CommerceMollie\Domain\Mollie\Command;
-
-use Backend\Modules\CommerceMollie\Domain\Mollie\MolliePaymentMethodDataTransferObject;
-
-class UpdateMolliePaymentMethod extends MolliePaymentMethodDataTransferObject
-{
-}
