+++ src/Backend/Modules/CommerceMollie/Domain/PaymentMethod/Command/UpdateMolliePaymentMethod.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Backend\Modules\CommerceMollie\Domain\PaymentMethod\Command;
+
+use Backend\Modules\CommerceMollie\Domain\PaymentMethod\MolliePaymentMethodDataTransferObject;
+
+class UpdateMolliePaymentMethod extends MolliePaymentMethodDataTransferObject
+{
+}
