+++ src/Backend/Modules/CommerceMollie/Domain/PaymentMethod/Command/UpdateMolliePaymentMethodHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Backend\Modules\CommerceMollie\Domain\Mollie\Command;
+namespace Backend\Modules\CommerceMollie\Domain\PaymentMethod\Command;
 
 use Backend\Modules\Commerce\Domain\PaymentMethod\Command\UpdatePaymentMethodHandler;
