+++ src/Backend/Modules/CommerceMollie/Resources/config/commands.yml
@@ -1,8 +1,8 @@
 services:
     commerce.handler.update_payment_method_mollie:
-        class: Backend\Modules\CommerceMollie\Domain\Mollie\Command\UpdateMolliePaymentMethodHandler
+        class: Backend\Modules\CommerceMollie\Domain\PaymentMethod\Command\UpdateMolliePaymentMethodHandler
         arguments:
             - "@commerce.repository.payment_method"
             - "@fork.settings"
         tags:
-            - { name: command_handler, handles: Backend\Modules\CommerceMollie\Domain\Mollie\Command\UpdateMolliePaymentMethod }
+            - { name: command_handler, handles: Backend\Modules\CommerceMollie\Domain\PaymentMethod\Command\UpdateMolliePaymentMethod }
